+++ src/components/Button.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import './Button.css';
+import '../style/Button.css';
 
 interface ButtonProps {
     onClick: React.MouseEventHandler<HTMLButtonElement>;
