+++ src/components/Home.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import './Home.css';
+import '../style/Home.css';
 
 const Home: React.FC = (): JSX.Element => {
     return (
